refactor: drop pandas for saving csv refactor: rename datagun to extracts
