+++ src/teamster/code_locations/dev/definitions.py
@@ -26,7 +26,7 @@ defs = Definitions(
             # adp,
             # airbyte,
             # amplify,
-            # datagun,
+            # extracts,
             # deanslist,
             # fivetran,
             # ldap,
@@ -44,7 +44,7 @@ defs = Definitions(
     #     *adp.schedules,
     #     *airbyte.schedules,
     #     *amplify.schedules,
-    #     *datagun.schedules,
+    #     *extracts.schedules,
     #     *fivetran.schedules,
     #     *ldap.schedules,
     #     *schoolmint.schedules,
