--- src/teamster/code_locations/kippcamden/datagun/__init__.py
@@ -1,7 +1,0 @@
-from teamster.code_locations.kippcamden.datagun.assets import assets
-from teamster.code_locations.kippcamden.datagun.schedules import schedules
-
-__all__ = [
-    "assets",
-    "schedules",
-]
