+++ src/teamster/code_locations/kippcamden/extracts/__init__.py
@@ -1,0 +1,7 @@
+from teamster.code_locations.kippcamden.extracts.assets import assets
+from teamster.code_locations.kippcamden.extracts.schedules import schedules
+
+__all__ = [
+    "assets",
+    "schedules",
+]
