+++ src/teamster/code_locations/kippcamden/extracts/jobs.py
@@ -1,7 +1,9 @@
 from dagster import define_asset_job
 
 from teamster.code_locations.kippcamden import CODE_LOCATION
-from teamster.code_locations.kippcamden.datagun.assets import powerschool_extract_assets
+from teamster.code_locations.kippcamden.extracts.assets import (
+    powerschool_extract_assets,
+)
 
 powerschool_extract_asset_job = define_asset_job(
     name=f"{CODE_LOCATION}_datagun_powerschool_extract_asset_job",
