--- src/teamster/code_locations/kippmiami/datagun/__init__.py
@@ -1,7 +1,0 @@
-from teamster.code_locations.kippmiami.datagun.assets import assets
-from teamster.code_locations.kippmiami.datagun.schedules import schedules
-
-__all__ = [
-    "assets",
-    "schedules",
-]
