+++ src/teamster/code_locations/kippmiami/extracts/schedules.py
@@ -1,7 +1,9 @@
 from dagster import ScheduleDefinition
 
 from teamster.code_locations.kippmiami import LOCAL_TIMEZONE
-from teamster.code_locations.kippmiami.datagun.jobs import powerschool_extract_asset_job
+from teamster.code_locations.kippmiami.extracts.jobs import (
+    powerschool_extract_asset_job,
+)
 
 powerschool_extract_assets_schedule = ScheduleDefinition(
     job=powerschool_extract_asset_job,
