--- src/teamster/code_locations/kippnewark/datagun/__init__.py
@@ -1,7 +1,0 @@
-from teamster.code_locations.kippnewark.datagun.assets import assets
-from teamster.code_locations.kippnewark.datagun.schedules import schedules
-
-__all__ = [
-    "assets",
-    "schedules",
-]
