--- src/teamster/code_locations/kipptaf/datagun/__init__.py
@@ -1,7 +1,0 @@
-from teamster.code_locations.kipptaf.datagun.assets import assets
-from teamster.code_locations.kipptaf.datagun.schedules import schedules
-
-__all__ = [
-    "assets",
-    "schedules",
-]
