+++ src/teamster/code_locations/kipptaf/definitions.py
@@ -16,8 +16,8 @@ from teamster.code_locations.kipptaf import (
     airbyte,
     amplify,
     asset_checks,
-    datagun,
     deanslist,
+    extracts,
     fivetran,
     ldap,
     overgrad,
@@ -52,7 +52,7 @@ defs = Definitions(
                 _google,
                 adp,
                 amplify,
-                datagun,
+                extracts,
                 deanslist,
                 ldap,
                 overgrad,
@@ -70,7 +70,7 @@ defs = Definitions(
         *adp.schedules,
         *airbyte.schedules,
         *amplify.schedules,
-        *datagun.schedules,
+        *extracts.schedules,
         *fivetran.schedules,
         *ldap.schedules,
         *schoolmint.schedules,
