+++ src/teamster/code_locations/kipptaf/extracts/__init__.py
@@ -1,0 +1,7 @@
+from teamster.code_locations.kipptaf.extracts.assets import assets
+from teamster.code_locations.kipptaf.extracts.schedules import schedules
+
+__all__ = [
+    "assets",
+    "schedules",
+]
