+++ src/teamster/code_locations/kipptaf/extracts/config/deanslist-annual.yaml
